+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntitySearchField.java
@@ -88,6 +88,7 @@ import static is.codion.swing.common.ui.component.Components.gridLayoutPanel;
 import static is.codion.swing.common.ui.component.Components.menu;
 import static is.codion.swing.common.ui.component.text.TextComponents.selectAllOnFocusGained;
 import static is.codion.swing.common.ui.layout.Layouts.borderLayout;
+import static is.codion.swing.framework.ui.component.EntitySearchField.SearchIndicator.WAIT_CURSOR;
 import static java.awt.event.InputEvent.CTRL_DOWN_MASK;
 import static java.awt.event.InputEvent.SHIFT_DOWN_MASK;
 import static java.awt.event.KeyEvent.*;
@@ -121,7 +122,7 @@ public final class EntitySearchField extends HintTextField {
    * Default value: {@link SearchIndicator#WAIT_CURSOR}
    */
   public static final PropertyValue<SearchIndicator> SEARCH_INDICATOR = Configuration.enumValue("is.codion.swing.framework.ui.component.EntitySearchField.searchIndicator",
-          SearchIndicator.class, SearchIndicator.WAIT_CURSOR);
+          SearchIndicator.class, WAIT_CURSOR);
 
   /**
    * The ways which a search field can indicate that a search is in progress.
@@ -143,12 +144,12 @@ public final class EntitySearchField extends HintTextField {
   private final State searchOnFocusLost = State.state(true);
   private final State searching = State.state();
   private final Value<Function<EntitySearchModel, Selector>> selectorFactory;
+  private final Value<SearchIndicator> searchIndicator;
 
   private SettingsPanel settingsPanel;
   private SingleSelectionValue singleSelectionValue;
   private MultiSelectionValue multiSelectionValue;
   private ProgressWorker<List<Entity>, ?> searchWorker;
-  private SearchIndicator searchIndicator = SEARCH_INDICATOR.get();
   private Consumer<Boolean> searchIndicatorListener;
 
   private Color backgroundColor;
@@ -167,7 +168,8 @@ public final class EntitySearchField extends HintTextField {
       TextComponents.lowerCase(getDocument());
     }
     searchOnFocusLost.set(builder.searchOnFocusLost);
-    searchIndicator(builder.searchIndicator);
+    searchIndicator = Value.value(builder.searchIndicator, builder.searchIndicator);
+    updateSearchIndicator();
     selectorFactory = Value.value(builder.selectorFactory, builder.selectorFactory);
     if (builder.selectAllOnFocusGained) {
       selectAllOnFocusGained(this);
@@ -218,12 +220,11 @@ public final class EntitySearchField extends HintTextField {
   }
 
   /**
-   * @param searchIndicator the search indicator type
+   * @return the value controlling the search indicator type
    * @see #SEARCH_INDICATOR
    */
-  public void searchIndicator(SearchIndicator searchIndicator) {
-    this.searchIndicator = requireNonNull(searchIndicator);
-    updateSearchIndicator();
+  public Value<SearchIndicator> searchIndicator() {
+    return searchIndicator;
   }
 
   /**
@@ -349,7 +350,7 @@ public final class EntitySearchField extends HintTextField {
     new SearchStringValue(this).link(model.searchString());
     model.searchString().addDataListener(searchString -> updateColors());
     model.entities().addListener(() -> setCaretPosition(0));
-    updateSearchIndicator();
+    searchIndicator.addListener(this::updateSearchIndicator);
     addFocusListener(new SearchFocusListener());
     addKeyListener(new EnterEscapeListener());
     linkToEnabledState(model.searchStringModified().not(), transferFocusAction, transferFocusBackwardAction);
@@ -364,7 +365,7 @@ public final class EntitySearchField extends HintTextField {
   }
 
   private Consumer<Boolean> createSearchIndicatorListener() {
-    switch (searchIndicator) {
+    switch (searchIndicator.get()) {
       case WAIT_CURSOR:
         return new WaitCursorWhileSearching();
       case PROGRESS_BAR: